File modifications

+++ libft/srcs/ft_numlen_base.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_numlen_base(int n, int base)
+int	ft_numlen_base(intmax_t n, int base)
 {
 	int	len;
